+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [Unreleased]
+
+### Improvements
+
+- Better MathML serialization of `\operatorname{}` and `\mathrm{}`
+
 ## 0.87.0 (2023-01-20)
 
 ### Improvements